requirements.txt
torch
torchvision
numpy
pillow
# Optional for DCT operations
opencv-python
scipy
